mise en place verification d'autorisation des pages backend membres et admin

+++ src/controller/BackendController.php
@@ -19,6 +19,7 @@ class BackendController
     private $sessionService;
     private $annonceDao;
     private $template;
+    private $messagingDao;
 
     function __construct()
     {
@@ -30,44 +31,15 @@ class BackendController
         $this->template->addExtension(new Twig_Extension_Debug());
     }
 
-    /** Permet de creer une nouvelle annonce*/
-    function addAnnonce($title, $content, $typeof, $tel, $email, $city, $author)
+    /** permet d'afficher la page d'inscription*/
+    function displayInscription()
     {
-        $affectedLines = $this->annonceDao->createAnnonce($title, $content, $typeof, $tel, $email, $city, $author);;
-
-        if ($affectedLines === false) {
-            throw new \Exception('Impossible d\'ajouter l\'annonce !');
-        } else {
-            echo $this->template->render('backend/my-annonces.html.twig');
-        }
+        echo $this->template->render('frontend/inscription.html.twig');
     }
 
-
-    /** permet de modifier une annonce existante */
-    function editAnnonce($annonceId, $template)
-    {
-        $annonceDao = new AnnonceDao();
-        $annonce = $annonceDao->getannonce($annonceId);
-        echo $this->template->render('backend/user-modify-annonce.html.twig', array('annonce' => $annonce));
-    }
-
-    /** permet valider les modifications d'une annonce existante */
-    function replaceAnnonce($id, $title, $content, $typeof, $tel, $email, $city)
-    {
-        $annonceDao = new AnnonceDao();
-        $affectedLines = $this->annonceDao->updateAnnonce($id, $title, $content, $typeof, $tel, $email, $city);
-        $annonces = $annonceDao->getMyAnnonces();
-        if ($affectedLines === false) {
-            throw new \Exception('Impossible de modifier l/annonce!');
-        } else {
-            header('location: index.php?action=displayPanelUser');
-        }
-    }
-
-    /** permet de renvoyer l'utilisateur à la page d'accueil si inscription avec succés*/
+    /** inscription + redirection à la page d'accueil si inscription avec succés*/
     function inscription($pseudo, $pass, $email)
     {
-
         $affectedLines = $this->memberDao->createMember($pseudo, password_hash($pass, PASSWORD_BCRYPT), $email);
         if ($affectedLines === false) {
             throw new \Exception('Tous les champs ne sont pas complétés');
@@ -76,92 +48,6 @@ class BackendController
             die();
         }
     }
-
-
-    /** Permet de se deconnecter */
-    function disconnect()
-    {
-        $this->sessionService->disconnect();
-        echo $this->template->render('frontend/connexion.html.twig');
-    }
-
-
-    /** permet d'afficher la page d'inscription*/
-    function displayInscription()
-    {
-
-        echo $this->template->render('frontend/inscription.html.twig');
-    }
-
-    /** permet d'afficher la page permettant de creer une nouvelle annonce*/
-    function displayNewAnnonce()
-    {
-
-        echo $this->template->render('backend/new-annonce-view.html.twig');
-
-    }
-
-    /** permet d'afficher la page contenant les annonces de l'utilisateur*/
-    function displayMyAnnonces()
-    {
-        echo $this->template->render('backend/my-annonces.html.twig');
-
-    }
-
-
-    /**annonces signalées*/
-    function listAnnoncesSpam()
-    {
-        $annonceDao = new AnnonceDao();
-        $annonces = $annonceDao->getSpamAnnonces();
-        echo $this->template->render('backend/spam.html.twig', array('annonces' => $annonces));
-    }
-
-    /**messages recus*/
-    function listMassages()
-    {
-        $messagingDao = new MessagingDao();
-        $messagings = $messagingDao->getAllMessages();
-        echo $this->template->render('backend/messages.html.twig', array('messagings' => $messagings));
-    }
-
-
-
-
-
-
-
-
-
-
-
-    /** annonce membre */
-    function annoncesMember()
-    {
-
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
     /** Permet de se connecter */
     function reqUser($mailconnect, $mdpconnect)
@@ -178,37 +64,6 @@ class BackendController
 
     }
 
-
-    /** permet d'afficher la page de connexion*/
-    function displayUserConnexion()
-    {
-        if ($this->sessionService->isClientAuthorized()) {
-            echo $this->template->render('backend/user-view.html.twig');
-        } else {
-            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
-            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
-        }
-    }
-
-
-    /** permet d'afficher la page principale de l'espace personnel de la personne identifié */
-    function displayUserPanel()
-    {
-        if ($this->sessionService->isClientAuthorized()) {
-            echo $this->template->render('backend/user-view.html.twig');
-        } else {
-            echo $this->template->render('frontend/connexion.html.twig');;
-        }
-    }
-
-
-
-
-
-
-
-
-    /*****************************************************************************/
     /** Permet de se connecter */
     function reqUserMaster($mailconnect, $mdpconnect)
     {
@@ -224,6 +79,33 @@ class BackendController
 
     }
 
+    /** Permet de se deconnecter */
+    function disconnect()
+    {
+        $this->sessionService->disconnect();
+        echo $this->template->render('frontend/connexion.html.twig');
+    }
+
+    /** permet d'afficher la page de connexion*/
+    function displayUserConnexion()
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            echo $this->template->render('backend/user-view.html.twig');
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /** permet d'afficher la page principale de l'espace personnel de la personne identifié */
+    function displayUserPanel()
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            echo $this->template->render('backend/user-view.html.twig');
+        } else {
+            echo $this->template->render('frontend/connexion.html.twig');;
+        }
+    }
 
     /** permet d'afficher la page de connexion*/
     function displayUserConnexionMaster()
@@ -236,7 +118,6 @@ class BackendController
         }
     }
 
-
     /** permet d'afficher la page principale de l'espace personnel de la personne identifié */
     function displayUserPanelMaster()
     {
@@ -247,22 +128,145 @@ class BackendController
         }
     }
 
-
-
-
-
-
-
-
-
-
-
-
-
-    /** permet de supprimer une annonce */
-    function deleteAdminPost($annonceId)
+    /** Permet de creer une nouvelle annonce*/
+    function addAnnonce($title, $content, $typeof, $tel, $email, $city, $author)
     {
-        {
+        if ($this->sessionService->isClientAuthorized()) {
+            $affectedLines = $this->annonceDao->createAnnonce($title, $content, $typeof, $tel, $email, $city, $author);;
+
+            if ($affectedLines === false) {
+                throw new \Exception('Impossible d\'ajouter l\'annonce !');
+            } else {
+                echo $this->template->render('backend/my-annonces.html.twig');
+            }
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /** permet de modifier une annonce existante */
+    function editAnnonce($annonceId)
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            $annonceDao = new AnnonceDao();
+            $annonce = $annonceDao->getannonce($annonceId);
+            echo $this->template->render('backend/user-modify-annonce.html.twig', array('annonce' => $annonce));
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /** permet valider les modifications d'une annonce existante */
+    function replaceAnnonce($id, $title, $content, $typeof, $tel, $email, $city)
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            $annonceDao = new AnnonceDao();
+            $affectedLines = $this->annonceDao->updateAnnonce($id, $title, $content, $typeof, $tel, $email, $city);
+            $annonces = $annonceDao->getMyAnnonces();
+            if ($affectedLines === false) {
+                throw new \Exception('Impossible de modifier l/annonce!');
+            } else {
+                header('location: index.php?action=displayPanelUser');
+            }
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /** permet d'afficher la page permettant de creer une nouvelle annonce*/
+    function displayNewAnnonce()
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            echo $this->template->render('backend/new-annonce-view.html.twig');
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /** permet d'afficher la page contenant les annonces de l'utilisateur*/
+    function displayMyAnnonces()
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            echo $this->template->render('backend/my-annonces.html.twig');
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/connexion.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /**annonces signalées*/
+    function listAnnoncesSpam()
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            $annonceDao = new AnnonceDao();
+            $annonces = $annonceDao->getSpamAnnonces();
+            echo $this->template->render('backend/spam.html.twig', array('annonces' => $annonces));
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/master.html.twig', ["hasFormError" => $hasFormError]);
+        }
+
+
+    }
+
+    /**messages recus*/
+    function listMessages($id)
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            $messagingDao = new MessagingDao();
+            $messagings = $messagingDao->getAllMessages();
+            $messaging = $messagingDao->getMessageById($id);
+            echo $this->template->render('backend/messages.html.twig', array('messagings' => $messagings, 'messaging' => $messaging));
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/master.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /********************************************************/
+    /** permet d'éditer le staus d'un message */
+    function editStatusMessage($id)
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            $messagings = $this->messagingDao->getAllMessages();
+            $messaging = $this->messagingDao->getMessageById($id);
+            echo $this->template->render('backend/message-status.html.twig', array('messaging' => $messaging, 'messagings' => $messagings));
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/master.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /** permet d'envoyer le mettre à jour le status d'un message */
+    function updateStatus($id, $status)
+    {
+        if ($this->sessionService->isClientAuthorized()) {
+            $messagings = $this->messagingDao->getAllMessages();
+            $affectedLines = $this->messagingDao->updateMessage($id, $status);
+            if ($affectedLines === false) {
+                throw new \Exception('Impossible de modifier le status !');
+            } else {
+                echo $this->template->render('backend/messages.html.twig', array('messagings' => $messagings));
+
+            }
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/master.html.twig', ["hasFormError" => $hasFormError]);
+        }
+    }
+
+    /*********************************************************************/
+
+
+    /** permet de supprimer une annonce signalée */
+    function deleteAdminAnnonce($annonceId)
+
+    {
+        if ($this->sessionService->isClientAuthorized()) {
             $annonceDao = new AnnonceDao();
             $affectedLines = $this->annonceDao->deleteAnnonce($annonceId);
             $annonces = $annonceDao->getSpamAnnonces();
@@ -273,17 +277,19 @@ class BackendController
                 echo $this->template->render('backend/spam.html.twig', array('annonces' => $annonces));
 
             }
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/master.html.twig', ["hasFormError" => $hasFormError]);
         }
     }
 
 
-
-
-    /** permet de supprimer une annonce */
+    /** permet de supprimer un message */
 
     function deleteMessage($messagingId)
+
     {
-        {
+        if ($this->sessionService->isClientAuthorized()) {
             $messagingDao = new MessagingDao();
             $affectedLines = $this->messagingDao->deleteMessage($messagingId);
             $messagings = $messagingDao->getAllMessages();
@@ -293,7 +299,10 @@ class BackendController
                 echo $this->template->render('backend/messages.html.twig', array('messagings' => $messagings));
 
             }
+        } else {
+            $hasFormError = isset($_GET['error']) && $_GET["error"]; // A faire passer depuis le routeur
+            echo $this->template->render('frontend/master.html.twig', ["hasFormError" => $hasFormError]);
         }
-    }
 
+    }
 }
